+++ CHANGELOG.md
@@ -19,7 +19,7 @@ Updates should follow the [Keep a CHANGELOG](http://keepachangelog.com/) princip
 - Nothing
 
 ### Fixed
-- Nothing
+- Could not create Date from timestamp when using locale with `,` as decimal separator.
 
 ### Security
 - Nothing